Refactor code in knowx/src/app/dashboard/page.tsx and knowx/src/app/api/phase1/route.ts
