+++ knowx/db/schema.ts
@@ -8,9 +8,6 @@ import {
   boolean,
   text,
 } from "drizzle-orm/pg-core";
-import { drizzle } from "drizzle-orm/postgres-js";
-import { eq } from "drizzle-orm";
-import postgres from "postgres";
 
 export const user = pgTable("user", {
   id: serial("id").primaryKey(),
@@ -45,105 +42,3 @@ export type User = typeof user.$inferSelect;
 export type SearchLog = typeof search_log.$inferSelect;
 export type LoginLog = typeof login_log.$inferSelect;
 export type ErrorLog = typeof error_log.$inferSelect;
-
-export async function getUserId({ newEmail }: { newEmail: string }) {
-  const queryClient = postgres(
-    `postgres://postgres:${process.env.DB_PASSWORD ?? ""}@0.0.0.0:5432/KnowX`
-  );
-  const db = drizzle(queryClient);
-
-  let recent = await db.select().from(user).where(eq(user.email, newEmail));
-  return recent[0].id;
-  // return -1;
-}
-
-export async function handleLogin({ profile }: any) {
-  try {
-    const queryClient = postgres(
-      `postgres://postgres:${process.env.DB_PASSWORD ?? ""}@0.0.0.0:5432/KnowX`
-    );
-    const db = drizzle(queryClient);
-
-    const uniqueUser: User[] = await db
-      .select()
-      .from(user)
-      .where(eq(user.email, profile.email));
-
-    if (uniqueUser.length === 0) {
-      const newUser = await db
-        .insert(user)
-        .values({
-          email: profile.email,
-          region: "MX",
-          createdAt: new Date().toISOString(),
-        })
-        .returning();
-
-      await db.insert(login_log).values({
-        idUser: newUser[0].id,
-        accessDate: new Date(),
-      });
-    } else {
-      await db.insert(login_log).values({
-        idUser: uniqueUser[0].id,
-        accessDate: new Date(),
-      });
-    }
-
-    return true;
-  } catch (error) {
-    await logError({
-      userId: -1,
-      description: String(error),
-      origin: "handleLogin",
-    });
-    return false;
-  }
-}
-
-export async function logSearch({
-  userId,
-  search,
-  feedback,
-}: {
-  userId: number;
-  search: string;
-  feedback: boolean;
-}) {
-  try {
-    const queryClient = postgres(
-      `postgres://postgres:${process.env.DB_PASSWORD ?? ""}@0.0.0.0:5432/KnowX`
-    );
-    const db = drizzle(queryClient);
-    await db.insert(search_log).values({
-      idUser: userId,
-      search,
-      timeOfSearch: new Date(),
-      feedback,
-    });
-  } catch (error) {
-    await logError({ userId, description: String(error), origin: "logSearch" });
-  }
-}
-
-async function logError({
-  userId,
-  description,
-  origin,
-}: {
-  userId: number;
-  description: string;
-  origin: string;
-}) {
-  const queryClient = postgres(
-    `postgres://postgres:${process.env.DB_PASSWORD ?? ""}@0.0.0.0:5432/KnowX`
-  );
-  const db = drizzle(queryClient);
-
-  await db.insert(error_log).values({
-    idUser: userId,
-    description,
-    errorTime: new Date(),
-    origin,
-  });
-}
